Move Jard commands to commands folder
