+++ lib/ruby_jard/commands/hide_command.rb
