+++ lib/ruby_jard/commands/jard_command.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require 'ruby_jard/commands/jard/show_command'
-require 'ruby_jard/commands/jard/hide_command'
-require 'ruby_jard/commands/jard/color_scheme_command'
-require 'ruby_jard/commands/jard/output_command'
-require 'ruby_jard/commands/jard/filter_command'
+require 'ruby_jard/commands/show_command'
+require 'ruby_jard/commands/hide_command'
+require 'ruby_jard/commands/color_scheme_command'
+require 'ruby_jard/commands/output_command'
+require 'ruby_jard/commands/filter_command'
 
 module RubyJard
   module Commands
