+++ lib/ruby_jard/commands/output_command.rb
