+++ lib/ruby_jard/commands/show_command.rb
